requirements.txt
python-multipart
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
pydantic-settings==2.4.0
Jinja2==3.1.4
aiofiles==23.2.1
redis==5.0.8
itsdangerous==2.2.0
httpx==0.27.0
pytest==8.3.2
starlette==0.38.2
pydantic[email]
bcrypt==4.0.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
